requirements.txt
streamlit==1.46.0
langchain==0.3.25
langchainhub==0.1.21
langchain-chroma==0.2.4
langchain_experimental==0.3.4
langchain-community==0.3.24
langchain_huggingface==0.2.0
langchain-groq==0.2.1
python-dotenv==1.1.0
pypdf==5.1.0
chromadb==0.5.23
sentence-transformers==3.3.1